Updated find_budget(...) signature
Since the calculate_cost(...) signature has changed, this needs to be updated.

+++ src/cost.py
@@ -14,7 +14,7 @@ def calculate_cost(test: Test) -> int:
     return 1
 
 
-def find_budget(objects: DataFrame, tests: list[Test], classes: set[str], cost_fn: Callable[[Test], float]) -> int:
+def find_budget(objects: DataFrame, tests: list[Test], classes: set[str], cost_fn: Callable[[Test], int]) -> int:
     """Implementation of the FindBudget procedure of the referenced paper"""
 
     def submod_f1(sub_tests: TestList):
